+++ python/cuml/dask/linear_model/linear_regression.py
@@ -80,9 +80,9 @@ class LinearRegression(BaseEstimator,
 
         Parameters
         ----------
-        X : dask cuDF dataframe (n_rows, n_features)
+        X : Dask cuDF dataframe  or CuPy backed Dask Array (n_rows, n_features)
             Features for regression
-        y : dask cuDF (n_rows, 1)
+        y : Dask cuDF dataframe  or CuPy backed Dask Array (n_rows, 1)
             Labels (outcome values)
         """
 
@@ -93,7 +93,7 @@ class LinearRegression(BaseEstimator,
         self.coef_ = self.local_model.coef_
         self.intercept_ = self.local_model.intercept_
 
-    def predict(self, X, delayed=True, parallelism=5):
+    def predict(self, X, delayed=True):
         """
         Make predictions for X and returns a dask collection.
 
@@ -103,19 +103,15 @@ class LinearRegression(BaseEstimator,
             Distributed dense matrix (floats or doubles) of shape
             (n_samples, n_features).
 
-        delayed : bool return lazy (delayed) result?
-
-        parallelism : int
-            Amount of concurrent partitions that will be processed
-            per worker. This bounds the total amount of temporary
-            workspace memory on the GPU that will need to be allocated
-            at any time.
+        delayed : bool (default = True)
+            Whether to do a lazy prediction (and return Delayed objects) or an
+            eagerly executed one.
 
         Returns
         -------
-        y : dask cuDF (n_rows, 1)
+        y : Dask cuDF dataframe  or CuPy backed Dask Array (n_rows, 1)
         """
-        return self._predict(X, delayed, parallelism)
+        return self._predict(X, delayed=delayed)
 
     def get_param_names(self):
         return list(self.kwargs.keys())