+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - PR #1750: LinearRegression CumlArray for configurable output
 - PR #1767: Single GPU decomposition models configurable output
 - PR #1778: Make cuML Handle picklable
+- PR #1738: cuml.dask refactor beginning and dask array input option for OLS, Ridge and KMeans
 
 ## Improvements
 - PR #1644: Add `predict_proba()` for FIL binary classifier